+++ releasenotes.md
@@ -4,7 +4,7 @@
 
 ### ✨ 功能更新
 
-- 添加对 Qihoo 360V6 的支持
+- 添加对 Qihoo 360V6、CMIOT AX18、ZN M2 的支持
 - 添加 bootconfig 命令，用于切换启动分区
 - 添加 untar 命令，用于解析 sysupgrade tar image
 - 调整相关代码，适配 NAND 机型
@@ -15,7 +15,9 @@
 
 ## 📡 支持设备
 
+- CMIOT AX18
 - Qihoo 360V6
+- ZN M2
 
 ## 📸 网页截图